JS-PROJECTS-DONE: Taskboard is online
Finished project using OOP, where every instance manipulates
it's own part of the DOM using inner properties of every class.
The project is divided on two folders: api and view, api stands
for inner logic(localStorage interactions), whereas view is
responsible for everything that user sees.
Each element on the page has it's own class, that make's it
easier to manipulate with the elements